+++ postgres/tests/test_statements.py
@@ -1377,7 +1377,7 @@ def test_load_pg_settings(aggregator, integration_check, dbm_instance, db_user):
     dbm_instance["dbname"] = "postgres"
     check = integration_check(dbm_instance)
     check._connect()
-    check.load_pg_settings(check.db)
+    check.load_pg_settings()
     if db_user == 'datadog_no_catalog':
         aggregator.assert_metric(
             "dd.postgres.error",
@@ -1400,9 +1400,14 @@ def test_pg_settings_caching(integration_check, dbm_instance):
     assert not check.pg_settings, "pg_settings should not have been initialized yet"
     check._connect()
     check.db_pool.get_main_db_pool()
+    # pg_settings is not loaded on connect
+    assert not check.pg_settings, "pg_settings should not have been initialized yet"
+    # pg_settings should now be lazy loaded
+    check.get_pg_settings()
+    assert check.pg_settings, "pg_settings should have been initialized"
     assert "track_activity_query_size" in check.pg_settings
     check.pg_settings["test_key"] = True
-    check.db_pool.get_main_db_pool()
+    check.get_pg_settings()
     assert (
         "test_key" in check.pg_settings
     ), "key should not have been blown away. If it was then pg_settings was not cached correctly"